Add files via upload
HR Analytics Project- Understanding the Attrition in HR

Attrition is critical in the industry 
these days. It’s the major problem which 
highlights in all the organizations. “Attrition is 
said to be the gradual reduction in the number 
of employees through retirement, resignation or 
death. It can also be said as Employee Turnover 
or Employee Defection” A well-trained and 
welladapted employee leaves the organization, it 
creates a vacuum. So, the organization loses key 
skills, knowledge and business relationships. 
Modern managers and personnel administrators 
are greatly interested in reducing Attrition in the 
organization, in such a way that it will 
contribute to the maximum effectiveness, 
growth, and progress of the organization. 
Therefore, we need an methods, algorithms to 
prediction of employee attrition using various 
data mining techniques .